requirements.txt
antlr4-python3-runtime==4.9.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
einops==0.7.0
filelock
Flask==3.0.0
fsspec==2023.12.2
future==1.0.0
gmpy2
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2
MarkupSafe
mpmath
networkx
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==23.2
pillow==10.2.0
pyglet==1.5.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
smplx==0.1.28
sympy
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.66.1
trimesh==4.1.8
triton==2.1.0
typing_extensions
urllib3==2.1.0
vit-pytorch==1.4.4
Werkzeug==3.0.1
zipp==3.17.0
